refactor(user): move repository to infra package
ui에 잘못 들어가있는 repository를 infra 패키지로 변경합니다.
